requirements.txt
asgiref==3.8.1
Django==5.0.6
Pillow==10.3.0
face-recognition==1.3.0
numpy==1.26.4
dlib==19.24.4
opencv-python==4.9.0.80
gunicorn==22.0.0  # Web server untuk production